ai-apis-and-standalone-apps/assignments/assignment-9/requirements.txt
gradio==4.44.0
Pillow==10.4.0
openai==1.47.0